TelegramBot/requirements.txt
pyserial==3.5
pyTelegramBotAPI==3.7.9